@@ -1,10 +1,11 @@
 module github.com/blocklessnetwork/b7s
 
-go 1.21.0
+go 1.22.0
 
 require (
 	github.com/Microsoft/go-winio v0.6.1
 	github.com/a-h/templ v0.2.648
+	github.com/armon/go-metrics v0.4.1
 	github.com/asaskevich/govalidator v0.0.0-20230301143203-a9d515a09cc2
 	github.com/blocklessnetwork/b7s-attributes v0.0.0
 	github.com/cavaliergopher/grab/v3 v3.0.1
