@@ -11,7 +11,12 @@ import (
 	"github.com/libp2p/go-libp2p/core/protocol"
 )
 
-func (r *Replica) send(to peer.ID, msg interface{}, protocol protocol.ID) error {
+func (r *Replica) send(ctx context.Context, to peer.ID, msg any, protocol protocol.ID) error {
+
+	ctx, span := r.tracer.Start(ctx, msgSendSpanName(msg, spanMessageSend))
+	defer span.End()
+
+	saveTraceContext(ctx, msg)
 
 	// Serialize the message.
 	payload, err := json.Marshal(msg)
@@ -20,7 +25,7 @@ func (r *Replica) send(to peer.ID, msg interface{}, protocol protocol.ID) error 
 	}
 
 	// We don't want to wait indefinitely.
-	ctx, cancel := context.WithTimeout(context.Background(), r.cfg.NetworkTimeout)
+	ctx, cancel := context.WithTimeout(ctx, r.cfg.NetworkTimeout)
 	defer cancel()
 
 	// Send message.
@@ -33,7 +38,12 @@ func (r *Replica) send(to peer.ID, msg interface{}, protocol protocol.ID) error 
 }
 
 // broadcast sends message to all peers in the replica set.
-func (r *Replica) broadcast(msg interface{}) error {
+func (r *Replica) broadcast(ctx context.Context, msg any) error {
+
+	ctx, span := r.tracer.Start(ctx, msgSendSpanName(msg, spanMessageBroadcast))
+	defer span.End()
+
+	saveTraceContext(ctx, msg)
 
 	// Serialize the message.
 	payload, err := json.Marshal(msg)
