@@ -1,15 +1,18 @@
 package pbft
 
 import (
+	"context"
 	"fmt"
 	"time"
 
+	"github.com/armon/go-metrics"
 	"github.com/libp2p/go-libp2p/core/peer"
 
 	"github.com/blocklessnetwork/b7s/models/blockless"
 	"github.com/blocklessnetwork/b7s/models/codes"
 	"github.com/blocklessnetwork/b7s/models/execute"
 	"github.com/blocklessnetwork/b7s/models/response"
+	"github.com/blocklessnetwork/b7s/telemetry/tracing"
 )
 
 // Execute fullfils the consensus interface by inserting the request into the pipeline.
@@ -26,7 +29,7 @@ func (r *Replica) Execute(client peer.ID, requestID string, timestamp time.Time,
 		Execute:   req,
 	}
 
-	err := r.processRequest(client, request)
+	err := r.processRequest(tracing.TraceContext(context.Background(), r.cfg.TraceInfo), client, request)
 	if err != nil {
 		return codes.Error, execute.Result{}, fmt.Errorf("could not process request: %w", err)
 	}
@@ -36,7 +39,7 @@ func (r *Replica) Execute(client peer.ID, requestID string, timestamp time.Time,
 }
 
 // execute executes the request AND sends the result back to origin.
-func (r *Replica) execute(view uint, sequence uint, digest string) error {
+func (r *Replica) execute(ctx context.Context, view uint, sequence uint, digest string) error {
 
 	// Sanity check, should not happen.
 	request, ok := r.requests[digest]
@@ -70,7 +73,7 @@ func (r *Replica) execute(view uint, sequence uint, digest string) error {
 
 	log.Info().Msg("executing request")
 
-	res, err := r.executor.ExecuteFunction(request.ID, request.Execute)
+	res, err := r.executor.ExecuteFunction(ctx, request.ID, request.Execute)
 	if err != nil {
 		log.Error().Err(err).Msg("execution failed")
 	}
@@ -93,8 +96,9 @@ func (r *Replica) execute(view uint, sequence uint, digest string) error {
 	}
 
 	msg := response.Execute{
-		Code:      res.Code,
-		RequestID: request.ID,
+		BaseMessage: blockless.BaseMessage{TraceInfo: r.cfg.TraceInfo},
+		Code:        res.Code,
+		RequestID:   request.ID,
 		Results: execute.ResultMap{
 			r.id: execute.NodeResult{
 				Result:   res,
@@ -121,10 +125,12 @@ func (r *Replica) execute(view uint, sequence uint, digest string) error {
 		return fmt.Errorf("could not sign execution request: %w", err)
 	}
 
-	err = r.send(request.Origin, msg, blockless.ProtocolID)
+	err = r.send(ctx, request.Origin, &msg, blockless.ProtocolID)
 	if err != nil {
 		return fmt.Errorf("could not send execution response to node (target: %s, request: %s): %w", request.Origin.String(), request.ID, err)
 	}
+
+	r.metrics.MeasureSinceWithLabels(pbftExecutionsTimeMetric, request.Timestamp, []metrics.Label{{Name: "function", Value: request.Execute.FunctionID}})
 
 	return nil
 }
