@@ -1,6 +1,7 @@
 package pbft
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/libp2p/go-libp2p/core/peer"
@@ -8,7 +9,7 @@ import (
 	"github.com/blocklessnetwork/b7s/models/blockless"
 )
 
-func (r *Replica) processRequest(from peer.ID, req Request) error {
+func (r *Replica) processRequest(ctx context.Context, from peer.ID, req Request) error {
 
 	pub, err := req.Origin.ExtractPublicKey()
 	if err != nil {
@@ -31,7 +32,7 @@ func (r *Replica) processRequest(from peer.ID, req Request) error {
 	if ok {
 		log.Info().Msg("request already executed, sending result to client")
 
-		err := r.send(req.Origin, result, blockless.ProtocolID)
+		err := r.send(ctx, req.Origin, &result, blockless.ProtocolID)
 		if err != nil {
 			return fmt.Errorf("could not send execution result back to client (request: %s, client: %s): %w", req.ID, req.Origin.String(), err)
 		}
@@ -66,7 +67,7 @@ func (r *Replica) processRequest(from peer.ID, req Request) error {
 	r.pending[digest] = req
 
 	// Broadcast a pre-prepare message.
-	err = r.sendPrePrepare(req)
+	err = r.sendPrePrepare(ctx, req)
 	if err != nil {
 		return fmt.Errorf("could not broadcast pre-prepare message (request: %v): %w", req.ID, err)
 	}
