@@ -1,6 +1,7 @@
 package executor
 
 import (
+	"github.com/armon/go-metrics"
 	"github.com/spf13/afero"
 
 	"github.com/blocklessnetwork/b7s/models/blockless"
@@ -18,12 +19,13 @@ var defaultConfig = Config{
 
 // Config represents the Executor configuration.
 type Config struct {
-	WorkDir         string   // directory where files needed for the execution are stored
-	RuntimeDir      string   // directory where the executable can be found
-	ExecutableName  string   // name for the executable
-	DriversRootPath string   // where are cgi drivers stored
-	FS              afero.Fs // FS accessor
-	Limiter         Limiter  // Resource limiter for executed processes
+	WorkDir         string           // directory where files needed for the execution are stored
+	RuntimeDir      string           // directory where the executable can be found
+	ExecutableName  string           // name for the executable
+	DriversRootPath string           // where are cgi drivers stored
+	FS              afero.Fs         // FS accessor
+	Limiter         Limiter          // Resource limiter for executed processes
+	Metrics         *metrics.Metrics // Metrics handle
 }
 
 type Option func(*Config)
@@ -60,5 +62,12 @@ func WithExecutableName(name string) Option {
 func WithLimiter(limiter Limiter) Option {
 	return func(cfg *Config) {
 		cfg.Limiter = limiter
+	}
+}
+
+// WithMetrics sets the metrics handler.
+func WithMetrics(metrics *metrics.Metrics) Option {
+	return func(cfg *Config) {
+		cfg.Metrics = metrics
 	}
 }
