@@ -6,6 +6,6 @@ import (
 )
 
 // cleanPath will return the path relative to the workdir.
-func (h *FStore) cleanPath(path string) string {
-	return filepath.Clean(strings.TrimPrefix(path, h.workdir))
+func (f *FStore) cleanPath(path string) string {
+	return filepath.Clean(strings.TrimPrefix(path, f.workdir))
 }
