@@ -1,6 +1,7 @@
 package fstore
 
 import (
+	"context"
 	"fmt"
 	"time"
 
@@ -8,9 +9,9 @@ import (
 )
 
 // Get retrieves a function manifest for the given function from storage.
-func (h *FStore) Get(cid string) (blockless.FunctionRecord, error) {
+func (f *FStore) Get(ctx context.Context, cid string) (blockless.FunctionRecord, error) {
 
-	fn, err := h.getFunction(cid)
+	fn, err := f.getFunction(ctx, cid)
 	if err != nil {
 		return blockless.FunctionRecord{}, fmt.Errorf("could not get function from store: %w", err)
 	}
@@ -18,30 +19,32 @@ func (h *FStore) Get(cid string) (blockless.FunctionRecord, error) {
 	return fn, nil
 }
 
-func (h *FStore) getFunction(cid string) (blockless.FunctionRecord, error) {
+func (f *FStore) getFunction(ctx context.Context, cid string) (blockless.FunctionRecord, error) {
 
-	function, err := h.store.RetrieveFunction(cid)
+	function, err := f.store.RetrieveFunction(ctx, cid)
 	if err != nil {
 		return blockless.FunctionRecord{}, fmt.Errorf("could not retrieve function record: %w", err)
 	}
 
-	// Update the "last retrieved" timestamp.
-	function.LastRetrieved = time.Now().UTC()
-	err = h.store.SaveFunction(function)
-	if err != nil {
-		h.log.Warn().Err(err).Str("cid", cid).Msg("could not update function record timestamp")
-	}
+	go func() {
+		// Update the "last retrieved" timestamp.
+		function.LastRetrieved = time.Now().UTC()
+		err = f.store.SaveFunction(context.Background(), function)
+		if err != nil {
+			f.log.Warn().Err(err).Str("cid", cid).Msg("could not update function record timestamp")
+		}
+	}()
 
 	return function, nil
 }
 
-func (h *FStore) saveFunction(fn blockless.FunctionRecord) error {
+func (f *FStore) saveFunction(ctx context.Context, fn blockless.FunctionRecord) error {
 
 	// Clean paths - make them relative to the current working directory.
-	fn.Archive = h.cleanPath(fn.Archive)
-	fn.Files = h.cleanPath(fn.Files)
-	fn.Manifest.Deployment.File = h.cleanPath(fn.Manifest.Deployment.File)
+	fn.Archive = f.cleanPath(fn.Archive)
+	fn.Files = f.cleanPath(fn.Files)
+	fn.Manifest.Deployment.File = f.cleanPath(fn.Manifest.Deployment.File)
 
 	fn.UpdatedAt = time.Now().UTC()
-	return h.store.SaveFunction(fn)
+	return f.store.SaveFunction(ctx, fn)
 }
