@@ -5,13 +5,16 @@ import (
 	"slices"
 	"sync"
 
+	"github.com/armon/go-metrics"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog"
 
 	"github.com/blocklessnetwork/b7s-attributes/attributes"
 	"github.com/blocklessnetwork/b7s/host"
+	"github.com/blocklessnetwork/b7s/info"
 	"github.com/blocklessnetwork/b7s/models/blockless"
 	"github.com/blocklessnetwork/b7s/node/internal/waitmap"
+	"github.com/blocklessnetwork/b7s/telemetry/tracing"
 )
 
 // Node is the entity that actually provides the main Blockless node functionality.
@@ -44,6 +47,10 @@ type Node struct {
 
 	executeResponses   *waitmap.WaitMap
 	consensusResponses *waitmap.WaitMap
+
+	// Telemetry
+	tracer  *tracing.Tracer
+	metrics *metrics.Metrics
 }
 
 // New creates a new Node.
@@ -69,7 +76,7 @@ func New(log zerolog.Logger, host *host.Host, store blockless.PeerStore, fstore 
 	n := &Node{
 		cfg: cfg,
 
-		log:      log.With().Str("component", "node").Logger(),
+		log:      log,
 		host:     host,
 		fstore:   fstore,
 		executor: cfg.Execute,
@@ -82,6 +89,9 @@ func New(log zerolog.Logger, host *host.Host, store blockless.PeerStore, fstore 
 		clusters:           make(map[string]consensusExecutor),
 		executeResponses:   waitmap.New(),
 		consensusResponses: waitmap.New(),
+
+		tracer:  tracing.NewTracer(tracerName),
+		metrics: metrics.Default(),
 	}
 
 	if cfg.LoadAttributes {
@@ -103,6 +113,12 @@ func New(log zerolog.Logger, host *host.Host, store blockless.PeerStore, fstore 
 	cn := newConnectionNotifee(log, store)
 	host.Network().Notify(cn)
 
+	n.metrics.SetGaugeWithLabels(nodeInfoMetric, 1, []metrics.Label{
+		{Name: "id", Value: n.ID()},
+		{Name: "version", Value: info.VcsVersion()},
+		{Name: "role", Value: n.cfg.Role.String()},
+	})
+
 	return n, nil
 }
 
@@ -111,13 +127,6 @@ func (n *Node) ID() string {
 	return n.host.ID().String()
 }
 
-func newRequestID() (string, error) {
-
-	// Generate a new request/executionID.
-	uuid, err := uuid.NewRandom()
-	if err != nil {
-		return "", fmt.Errorf("could not generate new request ID: %w", err)
-	}
-
-	return uuid.String(), nil
+func newRequestID() string {
+	return uuid.New().String()
 }
