@@ -1,24 +1,31 @@
 package node
 
 import (
+	"context"
+
 	"github.com/libp2p/go-libp2p/core/network"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/multiformats/go-multiaddr"
 	"github.com/rs/zerolog"
+	"go.opentelemetry.io/otel/trace"
 
 	"github.com/blocklessnetwork/b7s/models/blockless"
+	"github.com/blocklessnetwork/b7s/telemetry/b7ssemconv"
+	"github.com/blocklessnetwork/b7s/telemetry/tracing"
 )
 
 type connectionNotifiee struct {
-	log   zerolog.Logger
-	store blockless.PeerStore
+	log    zerolog.Logger
+	store  blockless.PeerStore
+	tracer *tracing.Tracer
 }
 
 func newConnectionNotifee(log zerolog.Logger, store blockless.PeerStore) *connectionNotifiee {
 
 	cn := connectionNotifiee{
-		log:   log.With().Str("component", "notifiee").Logger(),
-		store: store,
+		log:    log.With().Str("component", "notifiee").Logger(),
+		store:  store,
+		tracer: tracing.NewTracer("b7s.Notifiee"),
 	}
 
 	return &cn
@@ -26,13 +33,13 @@ func newConnectionNotifee(log zerolog.Logger, store blockless.PeerStore) *connec
 
 func (n *connectionNotifiee) Connected(network network.Network, conn network.Conn) {
 
+	ctx, span := n.tracer.Start(context.Background(), spanPeerConnected, connectionTraceOpts(conn)...)
+	defer span.End()
+
 	// Get peer information.
 	peerID := conn.RemotePeer()
 	maddr := conn.RemoteMultiaddr()
 	laddr := conn.LocalMultiaddr()
-
-	// We could save only the mutliaddress from which we receive this connection. However, we could theoretically have multiple connections
-	// and there's no reason to limit ourselves to a single address.
 
 	peer := blockless.Peer{
 		ID:        peerID,
@@ -56,13 +63,16 @@ func (n *connectionNotifiee) Connected(network network.Network, conn network.Con
 		Msg("peer connected")
 
 	// Store the peer info.
-	err := n.store.SavePeer(peer)
+	err := n.store.SavePeer(ctx, peer)
 	if err != nil {
 		n.log.Warn().Err(err).Str("id", peerID.String()).Msg("could not add peer to peerstore")
 	}
 }
 
 func (n *connectionNotifiee) Disconnected(_ network.Network, conn network.Conn) {
+
+	_, span := n.tracer.Start(context.Background(), spanPeerDisconnected, connectionTraceOpts(conn)...)
+	defer span.End()
 
 	maddr := conn.RemoteMultiaddr()
 	laddr := conn.LocalMultiaddr()
@@ -81,4 +91,14 @@ func (n *connectionNotifiee) Listen(_ network.Network, _ multiaddr.Multiaddr) {
 
 func (n *connectionNotifiee) ListenClose(_ network.Network, _ multiaddr.Multiaddr) {
 	// Noop
+}
+
+func connectionTraceOpts(conn network.Conn) []trace.SpanStartOption {
+	return []trace.SpanStartOption{
+		trace.WithAttributes(
+			b7ssemconv.PeerID.String(conn.RemotePeer().String()),
+			b7ssemconv.PeerMultiaddr.String(conn.RemoteMultiaddr().String()),
+			b7ssemconv.LocalMultiaddr.String(conn.LocalMultiaddr().String()),
+		),
+	}
 }
