@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/armon/go-metrics"
 	"github.com/libp2p/go-libp2p/core/peer"
 
 	"github.com/blocklessnetwork/b7s/consensus"
@@ -12,7 +13,6 @@ import (
 	"github.com/blocklessnetwork/b7s/models/codes"
 	"github.com/blocklessnetwork/b7s/models/execute"
 	"github.com/blocklessnetwork/b7s/models/request"
-	"github.com/blocklessnetwork/b7s/models/response"
 )
 
 func (n *Node) processRollCall(ctx context.Context, from peer.ID, req request.RollCall) error {
@@ -22,6 +22,8 @@ func (n *Node) processRollCall(ctx context.Context, from peer.ID, req request.Ro
 		n.log.Debug().Msg("skipping roll call as a non-worker node")
 		return nil
 	}
+
+	n.metrics.IncrCounterWithLabels(rollCallsSeenMetric, 1, []metrics.Label{{Name: "function", Value: req.FunctionID}})
 
 	log := n.log.With().Str("request", req.RequestID).Str("origin", req.Origin.String()).Str("function", req.FunctionID).Logger()
 	log.Debug().Msg("received roll call request")
@@ -46,17 +48,10 @@ func (n *Node) processRollCall(ctx context.Context, from peer.ID, req request.Ro
 		}
 	}
 
-	// Base response to return.
-	res := response.RollCall{
-		FunctionID: req.FunctionID,
-		RequestID:  req.RequestID,
-		Code:       codes.Error, // CodeError by default, changed if everything goes well.
-	}
-
 	// Check if we have this function installed.
-	installed, err := n.fstore.Installed(req.FunctionID)
+	installed, err := n.fstore.IsInstalled(req.FunctionID)
 	if err != nil {
-		sendErr := n.send(ctx, req.Origin, res)
+		sendErr := n.send(ctx, req.Origin, req.Response(codes.Error))
 		if sendErr != nil {
 			// Log send error but choose to return the original error.
 			log.Error().Err(sendErr).Str("to", req.Origin.String()).Msg("could not send response")
@@ -70,9 +65,9 @@ func (n *Node) processRollCall(ctx context.Context, from peer.ID, req request.Ro
 
 		log.Info().Msg("roll call but function not installed, installing now")
 
-		err = n.installFunction(req.FunctionID, manifestURLFromCID(req.FunctionID))
+		err = n.installFunction(ctx, req.FunctionID, manifestURLFromCID(req.FunctionID))
 		if err != nil {
-			sendErr := n.send(ctx, req.Origin, res)
+			sendErr := n.send(ctx, req.Origin, req.Response(codes.Error))
 			if sendErr != nil {
 				// Log send error but choose to return the original error.
 				log.Error().Err(sendErr).Str("to", req.Origin.String()).Msg("could not send response")
@@ -83,9 +78,10 @@ func (n *Node) processRollCall(ctx context.Context, from peer.ID, req request.Ro
 
 	log.Info().Str("origin", req.Origin.String()).Msg("reporting for roll call")
 
+	n.metrics.IncrCounterWithLabels(rollCallsAppliedMetric, 1, []metrics.Label{{Name: "function", Value: req.FunctionID}})
+
 	// Send positive response.
-	res.Code = codes.Accepted
-	err = n.send(ctx, req.Origin, res)
+	err = n.send(ctx, req.Origin, req.Response(codes.Accepted))
 	if err != nil {
 		return fmt.Errorf("could not send response: %w", err)
 	}
@@ -180,6 +176,8 @@ rollCallResponseLoop:
 // On successful issuance of the roll call request, we return the ID of the issued request.
 func (n *Node) publishRollCall(ctx context.Context, requestID string, functionID string, consensus consensus.Type, topic string, attributes *execute.Attributes) error {
 
+	n.metrics.IncrCounterWithLabels(rollCallsPublishedMetric, 1, []metrics.Label{{Name: "function", Value: functionID}})
+
 	// Create a roll call request.
 	rollCall := request.RollCall{
 		Origin:     n.host.ID(),
@@ -194,7 +192,7 @@ func (n *Node) publishRollCall(ctx context.Context, requestID string, functionID
 	}
 
 	// Publish the mssage.
-	err := n.publishToTopic(ctx, topic, rollCall)
+	err := n.publishToTopic(ctx, topic, &rollCall)
 	if err != nil {
 		return fmt.Errorf("could not publish to topic: %w", err)
 	}
