@@ -8,10 +8,12 @@ import (
 	"io"
 	"sync"
 
+	"github.com/armon/go-metrics"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/libp2p/go-libp2p/core/network"
 
 	"github.com/blocklessnetwork/b7s/models/blockless"
+	"github.com/blocklessnetwork/b7s/node/internal/pipeline"
 )
 
 // Run will start the main loop for the node.
@@ -28,7 +30,7 @@ func (n *Node) Run(ctx context.Context) error {
 	}
 
 	// Sync functions now in case they were removed from the storage.
-	err = n.fstore.Sync(false)
+	err = n.fstore.Sync(ctx, false)
 	if err != nil {
 		return fmt.Errorf("could not sync functions: %w", err)
 	}
@@ -97,10 +99,14 @@ func (n *Node) Run(ctx context.Context) error {
 					defer n.wg.Done()
 					defer func() { <-n.sema }()
 
-					err = n.processMessage(ctx, msg.ReceivedFrom, msg.GetData(), subscriptionPipeline)
+					n.metrics.IncrCounterWithLabels(topicMessagesMetric, 1, []metrics.Label{{Name: "topic", Value: name}})
+
+					err = n.processMessage(ctx, msg.ReceivedFrom, msg.GetData(), pipeline.PubSubPipeline(name))
 					if err != nil {
 						n.log.Error().Err(err).Str("id", msg.ID).Str("peer", msg.ReceivedFrom.String()).Msg("could not process message")
+						return
 					}
+
 				}(msg)
 			}
 		}(name, topic.subscription)
@@ -122,6 +128,8 @@ func (n *Node) listenDirectMessages(ctx context.Context) {
 
 		from := stream.Conn().RemotePeer()
 
+		n.metrics.IncrCounter(directMessagesMetric, 1)
+
 		buf := bufio.NewReader(stream)
 		msg, err := buf.ReadBytes('\n')
 		if err != nil && !errors.Is(err, io.EOF) {
@@ -132,9 +140,11 @@ func (n *Node) listenDirectMessages(ctx context.Context) {
 
 		n.log.Trace().Str("peer", from.String()).Msg("received direct message")
 
-		err = n.processMessage(ctx, from, msg, directMessagePipeline)
+		err = n.processMessage(ctx, from, msg, pipeline.DirectMessagePipeline())
 		if err != nil {
 			n.log.Error().Err(err).Str("peer", from.String()).Msg("could not process direct message")
+			return
 		}
+
 	})
 }
