@@ -1,13 +1,14 @@
 package store
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/cockroachdb/pebble"
 	"github.com/libp2p/go-libp2p/core/peer"
 )
 
-func (s *Store) RemovePeer(id peer.ID) error {
+func (s *Store) RemovePeer(_ context.Context, id peer.ID) error {
 
 	idBytes, err := id.MarshalBinary()
 	if err != nil {
@@ -23,7 +24,7 @@ func (s *Store) RemovePeer(id peer.ID) error {
 	return nil
 }
 
-func (s *Store) RemoveFunction(cid string) error {
+func (s *Store) RemoveFunction(_ context.Context, cid string) error {
 
 	key := encodeKey(PrefixFunction, cid)
 	err := s.remove(key)
