@@ -1,20 +1,24 @@
 package mocks
 
 import (
+	"context"
 	"testing"
 
+	"github.com/blocklessnetwork/b7s/models/blockless"
 	"github.com/blocklessnetwork/b7s/models/execute"
 )
 
+var _ (blockless.Executor) = (*Executor)(nil)
+
 type Executor struct {
-	ExecFunctionFunc func(string, execute.Request) (execute.Result, error)
+	ExecFunctionFunc func(context.Context, string, execute.Request) (execute.Result, error)
 }
 
 func BaselineExecutor(t *testing.T) *Executor {
 	t.Helper()
 
 	executor := Executor{
-		ExecFunctionFunc: func(string, execute.Request) (execute.Result, error) {
+		ExecFunctionFunc: func(context.Context, string, execute.Request) (execute.Result, error) {
 			return GenericExecutionResult, nil
 		},
 	}
@@ -22,6 +26,6 @@ func BaselineExecutor(t *testing.T) *Executor {
 	return &executor
 }
 
-func (e *Executor) ExecuteFunction(requestID string, req execute.Request) (execute.Result, error) {
-	return e.ExecFunctionFunc(requestID, req)
+func (e *Executor) ExecuteFunction(ctx context.Context, requestID string, req execute.Request) (execute.Result, error) {
+	return e.ExecFunctionFunc(ctx, requestID, req)
 }
