@@ -1,36 +1,27 @@
 package mocks
 
 import (
+	"context"
 	"testing"
-
-	"github.com/blocklessnetwork/b7s/models/blockless"
 )
 
 type FStore struct {
-	InstallFunc            func(string, string) error
-	GetFunc                func(string) (*blockless.FunctionManifest, error)
-	InstalledFunc          func(string) (bool, error)
-	InstalledFunctionsFunc func() ([]string, error)
-	SyncFunc               func(bool) error
+	InstallFunc     func(context.Context, string, string) error
+	IsInstalledFunc func(string) (bool, error)
+	SyncFunc        func(context.Context, bool) error
 }
 
 func BaselineFStore(t *testing.T) *FStore {
 	t.Helper()
 
 	fh := FStore{
-		GetFunc: func(string) (*blockless.FunctionManifest, error) {
-			return &GenericManifest, nil
-		},
-		InstallFunc: func(string, string) error {
+		InstallFunc: func(context.Context, string, string) error {
 			return nil
 		},
-		InstalledFunc: func(string) (bool, error) {
+		IsInstalledFunc: func(string) (bool, error) {
 			return true, nil
 		},
-		InstalledFunctionsFunc: func() ([]string, error) {
-			return nil, nil
-		},
-		SyncFunc: func(bool) error {
+		SyncFunc: func(context.Context, bool) error {
 			return nil
 		},
 	}
@@ -38,22 +29,14 @@ func BaselineFStore(t *testing.T) *FStore {
 	return &fh
 }
 
-func (f *FStore) Install(address string, cid string) error {
-	return f.InstallFunc(address, cid)
+func (f *FStore) Install(ctx context.Context, address string, cid string) error {
+	return f.InstallFunc(ctx, address, cid)
 }
 
-func (f *FStore) Get(cid string) (*blockless.FunctionManifest, error) {
-	return f.GetFunc(cid)
+func (f *FStore) IsInstalled(cid string) (bool, error) {
+	return f.IsInstalledFunc(cid)
 }
 
-func (f *FStore) Installed(cid string) (bool, error) {
-	return f.InstalledFunc(cid)
-}
-
-func (f *FStore) InstalledFunctions() ([]string, error) {
-	return f.InstalledFunctionsFunc()
-}
-
-func (f *FStore) Sync(haltOnError bool) error {
-	return f.SyncFunc(haltOnError)
+func (f *FStore) Sync(ctx context.Context, haltOnError bool) error {
+	return f.SyncFunc(ctx, haltOnError)
 }
