@@ -1,6 +1,7 @@
 package mocks
 
 import (
+	"context"
 	"testing"
 
 	"github.com/libp2p/go-libp2p/core/peer"
@@ -9,42 +10,44 @@ import (
 )
 
 type Store struct {
-	SavePeerFunc          func(blockless.Peer) error
-	SaveFunctionFunc      func(blockless.FunctionRecord) error
-	RetrievePeerFunc      func(peer.ID) (blockless.Peer, error)
-	RetrievePeersFunc     func() ([]blockless.Peer, error)
-	RetrieveFunctionFunc  func(string) (blockless.FunctionRecord, error)
-	RetrieveFunctionsFunc func() ([]blockless.FunctionRecord, error)
-	RemovePeerFunc        func(peer.ID) error
-	RemoveFunctionFunc    func(string) error
+	SavePeerFunc      func(context.Context, blockless.Peer) error
+	RetrievePeerFunc  func(context.Context, peer.ID) (blockless.Peer, error)
+	RetrievePeersFunc func(context.Context) ([]blockless.Peer, error)
+	RemovePeerFunc    func(context.Context, peer.ID) error
+
+	SaveFunctionFunc      func(context.Context, blockless.FunctionRecord) error
+	RetrieveFunctionFunc  func(context.Context, string) (blockless.FunctionRecord, error)
+	RetrieveFunctionsFunc func(context.Context) ([]blockless.FunctionRecord, error)
+	RemoveFunctionFunc    func(context.Context, string) error
 }
 
 func BaselineStore(t *testing.T) *Store {
 	t.Helper()
 
 	store := Store{
-		SavePeerFunc: func(blockless.Peer) error {
+		SavePeerFunc: func(context.Context, blockless.Peer) error {
 			return nil
 		},
-		SaveFunctionFunc: func(blockless.FunctionRecord) error {
-			return nil
-		},
-		RetrievePeerFunc: func(peer.ID) (blockless.Peer, error) {
+		RetrievePeerFunc: func(context.Context, peer.ID) (blockless.Peer, error) {
 			return GenericPeer, nil
 		},
-		RetrievePeersFunc: func() ([]blockless.Peer, error) {
+		RetrievePeersFunc: func(context.Context) ([]blockless.Peer, error) {
 			return []blockless.Peer{GenericPeer}, nil
 		},
-		RetrieveFunctionFunc: func(string) (blockless.FunctionRecord, error) {
-			return GenericFunctionRecord, nil
-		},
-		RetrieveFunctionsFunc: func() ([]blockless.FunctionRecord, error) {
-			return []blockless.FunctionRecord{GenericFunctionRecord}, nil
-		},
-		RemovePeerFunc: func(peer.ID) error {
+		RemovePeerFunc: func(context.Context, peer.ID) error {
 			return nil
 		},
-		RemoveFunctionFunc: func(string) error {
+
+		SaveFunctionFunc: func(context.Context, blockless.FunctionRecord) error {
+			return nil
+		},
+		RetrieveFunctionFunc: func(context.Context, string) (blockless.FunctionRecord, error) {
+			return GenericFunctionRecord, nil
+		},
+		RetrieveFunctionsFunc: func(context.Context) ([]blockless.FunctionRecord, error) {
+			return []blockless.FunctionRecord{GenericFunctionRecord}, nil
+		},
+		RemoveFunctionFunc: func(context.Context, string) error {
 			return nil
 		},
 	}
@@ -52,27 +55,27 @@ func BaselineStore(t *testing.T) *Store {
 	return &store
 }
 
-func (s *Store) SavePeer(peer blockless.Peer) error {
-	return s.SavePeerFunc(peer)
+func (s *Store) SavePeer(ctx context.Context, peer blockless.Peer) error {
+	return s.SavePeerFunc(ctx, peer)
 }
-func (s *Store) SaveFunction(function blockless.FunctionRecord) error {
-	return s.SaveFunctionFunc(function)
+func (s *Store) SaveFunction(ctx context.Context, function blockless.FunctionRecord) error {
+	return s.SaveFunctionFunc(ctx, function)
 }
-func (s *Store) RetrievePeer(id peer.ID) (blockless.Peer, error) {
-	return s.RetrievePeerFunc(id)
+func (s *Store) RetrievePeer(ctx context.Context, id peer.ID) (blockless.Peer, error) {
+	return s.RetrievePeerFunc(ctx, id)
 }
-func (s *Store) RetrievePeers() ([]blockless.Peer, error) {
-	return s.RetrievePeersFunc()
+func (s *Store) RetrievePeers(ctx context.Context) ([]blockless.Peer, error) {
+	return s.RetrievePeersFunc(ctx)
 }
-func (s *Store) RetrieveFunction(cid string) (blockless.FunctionRecord, error) {
-	return s.RetrieveFunctionFunc(cid)
+func (s *Store) RetrieveFunction(ctx context.Context, cid string) (blockless.FunctionRecord, error) {
+	return s.RetrieveFunctionFunc(ctx, cid)
 }
-func (s *Store) RetrieveFunctions() ([]blockless.FunctionRecord, error) {
-	return s.RetrieveFunctionsFunc()
+func (s *Store) RetrieveFunctions(ctx context.Context) ([]blockless.FunctionRecord, error) {
+	return s.RetrieveFunctionsFunc(ctx)
 }
-func (s *Store) RemovePeer(id peer.ID) error {
-	return s.RemovePeerFunc(id)
+func (s *Store) RemovePeer(ctx context.Context, id peer.ID) error {
+	return s.RemovePeerFunc(ctx, id)
 }
-func (s *Store) RemoveFunction(id string) error {
-	return s.RemoveFunctionFunc(id)
+func (s *Store) RemoveFunction(ctx context.Context, id string) error {
+	return s.RemoveFunctionFunc(ctx, id)
 }
